Refactor Environment Variable Testing: Uses Getter to make it more readable
The getter has to use the pointer notation *Configuration so the passed method gets evaluated with the state that is initialized later in the test. Otherwise it would be always yield the default (in case of MobDoneSquash: true).
As the expectEnvironmentVariableSetsConfiguration method uses interface{} which is generic and can be used for all sorts of types of Configuration properties, the boolean has to be "unpacked" to the interface{} type first. It's like a cast and uses a small helper function

One Downside remains: the Test is now referencing the global configuration, rather than the one returned by setEnv.
Which means this test now only works as long as configuration stays a global.
The lambda way, which was less readable, would have worked even if the configuration was only locally defined.