Speed up hasCommits check
`git rev-list --all` is very slow for large repositories
(order of 5s on a repo with 500k commits).

This alternative version works by checking whether HEAD points at a valid commit,
by using a git command that returns an error code if not.
This means it has to use the 'silentgitignorefailure'
version of the git invocation functions, and check for an error.